Add frontend and finish tutorial

+++ app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from app.schemas import PostCreate, PostResponse, UserRead, UserCreate, UserUpdate
-from app.db import Post, create_db_and_tables, get_async_session
+from app.db import Post, create_db_and_tables, get_async_session, User
 from app.images import imagekit
 from imagekitio.models.UploadFileRequestOptions import UploadFileRequestOptions
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -29,6 +29,7 @@ app.include_router(fastapi_users.get_users_router(UserRead, UserUpdate), prefix=
 async def upload_file(
     file   : UploadFile = File(...),
     caption: str = Form(""),
+    user   : User = Depends(current_active_user),
     session: AsyncSession = Depends(get_async_session)
     ):
     temp_file_path = None
@@ -44,6 +45,7 @@ async def upload_file(
 
         if upload_result.response_metadata.http_status_code == 200:
             post = Post(
+                user_id   = user.id,
                 caption   = caption,
                 url       = upload_result.url,
                 file_type = "video" if file.content_type.startswith("video/") else "image",
@@ -62,28 +64,36 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
-    session: AsyncSession = Depends(get_async_session)
+    session: AsyncSession = Depends(get_async_session),
+    user   : User = Depends(current_active_user),
     ):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts  = [row[0] for row in result.all()]
+
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id: u.email for u in users}
 
     posts_data = []
     for post in posts:
         posts_data.append(
             {
                 "id": str(post.id),
+                "user_id": str(post.user_id),
                 "caption": post.caption,
                 "url": post.url,
                 "file_type": post.file_type,
                 "file_name": post.file_name,
-                "created_at": post.created_at.isoformat()
+                "created_at": post.created_at.isoformat(),
+                "is_owner": post.user_id == user.id,
+                "email": user_dict.get(post.user_id, "Unknown")
             }
         )
 
     return {"posts": posts_data}
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session)):
+async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session), user: User = Depends(current_active_user)):
     try:
         post_uuid = uuid.UUID(post_id)
 
@@ -93,6 +103,9 @@ async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_se
 
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
+        
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="You don't have premission to delete this post")
         
         await session.delete(post)
         await session.commit()
